% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donations_scrape.R
\name{donations_scrape}
\alias{donations_scrape}
\title{donations_scrape}
\usage{
donations_scrape(inds, years, chambers = c("H", "S"), member = "Y")
}
\arguments{
\item{inds}{vector of industries in string format to scrape donations from}

\item{years}{vector of election years in string format identifying the elections to scrape data from}

\item{chambers}{vector of chambers of Congress in string format to scrape donations to, Default: c("H", "S")}

\item{member}{either "Y" or "N", include only donations to members of Congress or donations to members and failed candidate, Default: 'Y'}
}
\value{
outputs a dataframe containing contribution data including candidate information, the amount donated, the chamber of congress, the election year, and the donating industry
}
\description{
scrapes OpenSecrets political campaign contribution data. This function collects donations by specific industries to congresspeople and congressional candidates
}
\details{
very large requests will result in overtaxing the OpenSecrets servers and causing a HTTP 429 error, wait and rerun smaller requests when this error occurs
}
\examples{
\dontrun{
if(interactive()){
 donations_scrape("", "2004", "H", "N")
 }
}
}
