% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretscrape-package.R
\docType{package}
\name{secretscrape-package}
\alias{secretscrape}
\alias{secretscrape-package}
\title{secretscrape: Scrape opensecrets.org For Congressional Political Donation Data}
\description{
This package can be used to obtain data on industry specific political donations to Congress from opensecrets.org. This data is available at the individual Congressional candidate level from 1990-2020. For example, using this package allows you to obtain the agriculutural industry's donations to House and Senate candidates in 2010.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/username/secretscrape/isssues}
}

}
\author{
\strong{Maintainer}: Harrison Karp \email{hkarp@wesleyan.edu}

}
\keyword{internal}
